+++ internal/service/wallets.go
@@ -8,9 +8,9 @@ import (
 )
 
 type WalletsService interface {
-	Create(userID int, name string, balance float64, timestamp *time.Time, token string) (*domain.WalletDTO, error)
-	GetAll(userID int) (*[]domain.WalletDTO, error)
-	GetByName(name string, userID int) (*domain.WalletDTO, error)
+	Create(userID int64, name string, balance float64, timestamp *time.Time, token string) (*domain.WalletDTO, error)
+	GetAll(userID int64) (*[]domain.WalletDTO, error)
+	GetByName(name string, userID int64) (*domain.WalletDTO, error)
 }
 type walletsService struct {
 	repo repository.WalletsRepository
@@ -21,7 +21,7 @@ func NewWalletsService(repo repository.WalletsRepository) WalletsService {
 		repo: repo,
 	}
 }
-func (s *walletsService) Create(userID int, name string, balance float64, timestamp *time.Time, token string) (*domain.WalletDTO, error) {
+func (s *walletsService) Create(userID int64, name string, balance float64, timestamp *time.Time, token string) (*domain.WalletDTO, error) {
 	walletDTO := domain.WalletDTO{
 		UserID:    userID,
 		Name:      name,
@@ -30,14 +30,14 @@ func (s *walletsService) Create(userID int, name string, balance float64, timest
 	}
 	return s.repo.Create(&walletDTO, token)
 }
-func (s *walletsService) GetAll(userID int) (*[]domain.WalletDTO, error) {
+func (s *walletsService) GetAll(userID int64) (*[]domain.WalletDTO, error) {
 	token := jwt.GenerateToken(nil, &jwt.Payload{
 		Subject: userID,
 	})
 
 	return s.repo.GetAll(token)
 }
-func (s *walletsService) GetByName(name string, userID int) (*domain.WalletDTO, error) {
+func (s *walletsService) GetByName(name string, userID int64) (*domain.WalletDTO, error) {
 	token := jwt.GenerateToken(nil, &jwt.Payload{
 		Subject: userID,
 	})
